+++ src/main/java/tomcom/kartGame/scenes/menus/SplashScreen.java
@@ -1,4 +1,4 @@
-package tomcom.kartGame.scenes;
+package tomcom.kartGame.scenes.menus;
 
 import tomcom.kartGame.game.GameMain;
